Added filepath option to cli and gui
I had to refactor a bit how filepaths were handled as this was confusing. We now just save the filepath in the config class and use this to work out where to save files. Previously we could not specify the filepath for the main app and it was assumed to be cwd. However if you manually call some of the Result methods you could specify a filepath, I have removed that in favour of just using the Config class. I have added a filebrowser to the gui to select the filepath. I have also fixed the Results tests, but others are failing