Improve how ca ports are configured
Previously we attempted to set the CA addr and repeater ports
from within python. The preferred way is to read the current
values from the users environment. We now do this and raise
an error or log a warning if the variables are incompatible
with the args supplied to the app.