Improve handling of filepath/name in Results class
Each instance of the results class now stores its own filename and
filepath which is the only filename/filepath used to save data. These
are either set from a passed Config object or set directly when loading
a csv file. This allows us to create an initial results object and
directory and also later create a new Results object and directory in a
new location with a new name using the same code.